getSELinuxfs: fix/improve/speedup
I have noticed that libselinux (the C library) implementation tries
the default path (/sys/fs/selinux) first (checking its fstype and
readonly flag), then the old default (/selinux), and only if that
fails it goes on to parse /proc/self/mountinfo in order to find
the selinux mount point. The relevant commit is
https://github.com/SELinuxProject/selinux/commit/e3cab998b48ab

This commit does the same for the Go implementation, in particular:

1. Modify getSELinuxfs() to avoid parsing mountinfo if at all
possible.

Try the default path (/sys/fs/selinux) first, checking its
filesystem type and whether it is read only. In most cases it
should be enough.

If that fails, go the slow path of parsing /proc/self/mountinfo,
but check /proc/filesystems filed to see if selinuxfs is
available and so it makes sense to proceed.

2. Fix, simplify, and improve mountinfo parsing:
 - simplify search for selinuxfs;
 - use splitN to avoid excessive splitting;
 - in case of multiple selinux mounts, don't take the first one
   and give up if it is readonly; keep going.

3. Treat inability to open /proc files as fatal (as the code
is supposed to be run once and then just return what was
obtained during the first run).

4. As a side effect, fix the read only bit checking (which
only worked for stRdOnly==1), and make the check more strict
(check for fstype).

5. Add a unit test for mountinfo parsing function.

In practice, this should not change anything but speed up
the first call to the library.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>