go-selinux: Use selinux build tag in go-selinux/
Now that we have a stub implementation for !linux, give Linux users
access to that stub when they don't set the 'selinux' build tag.  For
example, they may have a kernel without SELinux compiled in.  Or they
may, for whatever reason, want a particular tool to build without real
SELinux support.

The syntax is described in [1].  In order to get the real
implementation, you must be on Linux and you must have
selinux set.

This unwinds e934c8fb (selinux_test.go: remove selinux build tag,
2018-03-30, #26), now that the real implementation depends on the
selinux build tag.  This commit does not address the "slightly easier
to run tests from the command line" point raised in that commit.

I've also update .travis.yml to excercise both the real and stub
implementations, although the stub testing is very minimal.  It would
be nice if there was an interface for the whole API so we could ensure
that both the real and stub implementations provided the same
interface, but that's too big a change for me to bite off here.

[1]: https://golang.org/pkg/go/build/#hdr-Build_Constraints

Signed-off-by: W. Trevor King <wking@tremily.us>