Release SELinux labels on error in init function.
Currently if we have an error in the options, we will leak and allocated
SELinux label.  This change will release on error.

Also return the mountlabel, even if the user specifies to disable SELinux separation.  This keeps other containers from being able to look at this privileged
containers content, as well makeing sure all content has a label.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>