Need to set process attributes not task
The current setting of the kernel is failing to set the kernel keyring
within runc.  Changing to setting the process field instead of the
thread field seems to have fixed the issue.  With this change runc
is labeling the kernel keyring correctly so that it can be used within
the container.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>