Fix [Set]EnforceMode and SecurityCheckContext
While working on minor improvements to tests, I found out
that SetEnforceMode() is not working even for the root user:

> --- FAIL: TestSetEnforceMode (0.00s)
>    selinux_linux_test.go:94: Enforcing Mode:  -1
>    selinux_linux_test.go:96: Default Enforce Mode:  1
>    selinux_linux_test.go:100: setting selinux mode to enforcing failed: /sys/fs/selinux/enforce not on procfs

If you look closer, you'll see that EnforceMode() is not working either,
since it returns -1.

The problem is, these functions, as well as SecurityCheckContext(),
are using readCon() and writeCon() methods, which require the files
being read/written to be on `procfs`.

In these cases, though, the files are on `selinuxfs`, and the filesystem
check is not really required since we already checked that during
selinuxfs mount point search.

So, just use ioutil.ReadFile/WriteFile here.

While at it
- convert code to use path.Join() instead of fmt.Sprintf()
- add a test case for SecurityCheckContext

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>