Chcon(): use parallelized Walk()
This uses the code from
https://github.com/stretchr/powerwalk/blob/master/walker.go
(which is under the MIT license, i.e. compatible with Apache).

Unfortunately, stretchr/powerwalk is based on filepath.Walk(),
which does stat() on every entry, so it can't be used directly,
and had to be modified to use karrick/godirwalk.Walk.

Benchmarking shows at least 2x improvement over the previous
commit, and about 3x over the original implementation with
filepath.Walk().

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>