Makefile: fix vendor target
Commit bb3d708b added this target. Alas, the `go mod tidy` never
worked, as it was written as part of `export` statement:

	export GO111MODULE=on \
		$(GO) mod tidy && \
		...

which is the same as

	export GO111MODULE=on $(GO) mod tidy && ...

which exports a bunch of variables, such as `go`, `mod`, and `tidy`
(or would export them if they existed), but does not run `go mod tidy`.

The fix would be to add a semicolon after the `export` statement,
but since exporting is not really required here, let's just
remove the `export` keyword.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>