travis: reorganize, add ppc to the matrix
This way, we will test on both amd64 and ppc64.

Now, git-validation and build-cross are separate things that do no make
sense to have in the matrix, so separate them out to additional jobs.

While at it, allow go: tip to fail.

NOTE that testing with and without selinux tag is handled by the
Makefile, so we do not add this dimension to the matrix.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>