Free SELinux labels when the labels are not being used
This is a cherry-pick of the following from projectatomic/docker altered
for this version of opencontainers/selinux.

https://github.com/projectatomic/docker/pull/381

Currently when lots of containers are created as disabled, each
container will leak and MCS Label.  Eventually the system will run out
of labels and go into an infinate loop looking for labels.

This also fixes a problem that caused relabeling to occur even when
label=disable was specified, such as with privileged containers.

Signed-off-by: Robb Manes <rmanes@redhat.com>