Use native error wrapping
The main benefit of using pkg/errors was the ability to wrap errors.
(The other benefit is saving the error traceback but I doubt any users
of go-selinux rely on that).

However, since Go 1.13 one can use native error wrapping and unwrapping
(fmt.Errorf with %w, errors.Is and errors.As).

Replace all uses of pkg/errors with stdlib. For wrapping errors, use
either %w or, where appropriate, os.PathError.

This change can be disruptive to users who rely on wrapped errors.
They should change to using errors.Is / errors.As to unwrap to check the
error type.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>