Chcon,Relabel: do process the root entry last
Commit 8501cfdf does not guarantee that the root entry is processed
last -- due to parallel nature of the pwalk it one of the last to be
processed, but might be not the last one.

Rewrite the code to process the root entry after all the others
are done.

Remove the test for root entry being last as the last = p assignment
is racy, and this is now very obvious from the code that the root entry
is indeed the last one.

Fixes: 8501cfdfaeec1072e7e
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>