pkg/pwalk: deprecate in favor of pkg/pwalkdir
The package was already deprecated in the README, but not in code.
This patch marks the code as deprecated, as go1.15 reached EOL, so
all consumers should be able to use the new package.

In a follow-up, we remove it (after it's been in a release to allow
users to migrate).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>