remove vendor directory
Using a vendor directory is common for repositories that produce binaries,
but usually not for library-only repositories. Having a vendor directory
can be useful to review changes in dependencies, but as this repository
only has a single dependency (which is well-known and (usually) requires
minimal review), it's probably ok to drop the vendor directory.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>