golangci: show all linting issues
By default, output is limited to 50 issues per linter, and 3 "same"
issues. When enabling a new linter, this requires multiple runs of
the linter to get a list of all issues.

Let's assume we don't introduce "many" new issues at a time, and just
show all the issues that were found.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>