Add a TODO to remove min/max
Go 1.21 has built-in min and max, so there is no need to define our own
(OTOH it's not a problem to have it).

I tried to move min/max to a separate file having "//go:build !go1.21"
build constraint, but go1.21.0 complains:

> max requires go1.21 or later (-lang was set to go1.19; check go.mod)
> min requires go1.21 or later (-lang was set to go1.19; check go.mod)

I have no idea how to work around that; so let's just add a TODO item to
remove min/max when it's time.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>