SetKeyLabel: add thread group leader requirement
Since commit bbbc51cbc717 ("Need to set process attributes not task")
SetKeyLabel and KeyLabel operate on /proc/self/attr/keycreate, which
can only be modified by the thread group leader; a non thread group
leader thread will get EACCES:

> write /proc/self/attr/keycreate: permission denied

Let's document that, and return a more meaningful error (ErrNotTGLeader)
if that is the case.

Modify the test case accordingly, i.e. skip it if the current thread is
not the thread group leader.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>